Authentication implementation

+++ App/Backend/Modules/Posts/PostsController.php
@@ -9,28 +9,65 @@ use \Entity\Comment;
 use \FormBuilder\CommentFormBuilder;
 use \FormBuilder\NewsFormBuilder;
 use \OCFram\FormHandler;
+use OCFram\Resize;
 
 class PostsController extends BackController
 {
+    use Resize;
+
     public function executeDelete(HTTPRequest $request)
     {
-        $postId = $request->getData('id');
+        $requestToken = $request->getData('token');
+        $userToken =  $this->app()->user()->getAttribute('token');
+        if ($requestToken == $userToken) {
+            $postId = $request->getData('id');
 
-        $this->managers->getManagerOf('Post')->delete($postId);
-        //$this->managers->getManagerOf('Comment')->deleteFromNews($postId);
+            $this->managers->getManagerOf('Post')->delete($postId);
+            $this->managers->getManagerOf('Comment')->deleteFromNews($postId);
 
-        $this->app->user()->setFlash('L\'article a bien été supprimé !', "alert alert-success");
+            $this->app->user()->setFlash('L\'article a bien été supprimé !', "alert alert-success");
 
-        $this->app->httpResponse()->redirect('.');
+            $this->app->httpResponse()->redirect('.');
+        } else {
+            $this->app->user()->setFlash('Les tokens ne correspondent pas !', "alert alert-danger");
+            $this->app->httpResponse()->redirect('.');
+        }
     }
 
     public function executeDeleteComment(HTTPRequest $request)
     {
-        $this->managers->getManagerOf('Comment')->delete($request->getData('id'));
+        $requestToken = $request->getData('token');
+        $userToken =  $this->app()->user()->getAttribute('token');
 
-        $this->app->user()->setFlash('Le commentaire a bien été supprimé !', "alert alert-success");
+        if ($requestToken == $userToken) {
+            $idComment = $request->getData('id');
 
-        $this->app->httpResponse()->redirect('.');
+            $idArticle = $this->managers->getManagerOf('Comment')->getIdArticle($idComment);
+
+            $this->managers->getManagerOf('Comment')->deleteWithChildren($idComment);
+
+            $this->app->user()->setFlash('Le commentaire a bien été supprimé !', "alert alert-success");
+
+            if (preg_match('/admin\/comments/', $_SERVER['HTTP_REFERER'])) {
+                $post = $this->managers->getManagerOf('Post')->getUnique($idArticle);
+
+                $url = $this->app->router()->generateUri("invalidComments");
+
+                $this->app->httpResponse()->redirect($url);
+            }
+
+            $post = $this->managers->getManagerOf('Post')->getUnique($idArticle);
+
+            $url = $this->app->router()->generateUri("showPost", [$post->slug(), $idArticle]);
+
+            $this->app->httpResponse()->redirect($url . '#form');
+        } else {
+            $this->app->user()->setFlash('Les tokens ne correspondent pas !', "alert alert-danger");
+            if (preg_match('/\/blog\//', $_SERVER['HTTP_REFERER'])) {
+                $this->app->httpResponse()->redirect($_SERVER['HTTP_REFERER'].'#form-comment');
+            }
+            $this->app->httpResponse()->redirect($_SERVER['HTTP_REFERER']);
+        }
     }
 
     public function executeIndex(HTTPRequest $request)
@@ -38,6 +75,8 @@ class PostsController extends BackController
         $this->page->addVar('title', 'Gestion des articles');
 
         $manager = $this->managers->getManagerOf('Post');
+
+        $invalidComments = $this->managers->getManagerOf('Comment')->countInvalidComments();
 
         if ($request->getExists("page") and $request->getData("page") == 1) {
             $this->app->httpResponse()->redirect("/admin/");
@@ -53,10 +92,16 @@ class PostsController extends BackController
 
         foreach ($listePosts->getCurrentPageResults() as $post) {
             $author = $usersManager->getUnique($post->idAuteur());
-            $post->setAuteur($author->nom() . " " . $author->prenom() . " (pseudo : " . $author->pseudo() . ")");
+            if ($author) {
+                $post->setAuteur($author->nom() . " " . $author->prenom() . " (pseudo : " . $author->pseudo() . ")");
+            }
+            $postUri = $this->app()->router()->generateUri('showPost', [$post->slug(),$post->id()]);
+            $postUri = '<a href="'.$postUri.'">'.$post->titre().'</a>';
+            $post->postUri = $postUri;
         }
 
         $this->page->addVar('listePosts', $listePosts);
+        $this->page->addVar('invalidComments', $invalidComments);
 
         $this->addView();
     }
@@ -98,7 +143,7 @@ class PostsController extends BackController
 
         $form = $formBuilder->form();
 
-        $formHandler = new FormHandler($form, $this->managers->getManagerOf('Comment'), $request);
+        $formHandler = new FormHandler($form, $this->managers->getManagerOf('Comment'), $request, $this->app->user());
 
         if ($formHandler->process()) {
             $this->app->user()->setFlash('Le commentaire a bien été modifié', "alert alert-success");
@@ -110,36 +155,13 @@ class PostsController extends BackController
     }
 
     /**
-     * Resize the $source image. The result will be $uploadfile
-     * The php gd extension must be enabled
-     * @param string $source : path to the source image
-     * @param mixed $uploadfile : path where the resized image will be placed
-     * @param int $width : width of the resized image
-     * @param int $height : height of the resized image
-     *
-     * @return void
-     */
-    private function resize(string $source, $uploadfile, int $width, int $height): void
-    {
-        $source = imagecreatefromjpeg($source);
-        $destination = imagecreatetruecolor($width, $height);
-
-        $largeur_source = imagesx($source);
-        $hauteur_source = imagesy($source);
-
-        imagecopyresampled($destination, $source, 0, 0, 0, 0, $width, $height, $largeur_source, $hauteur_source);
-
-        imagejpeg($destination, $uploadfile);
-    }
-
-    /**
      * @param Post|null $post=null
      *
      * @return array
      * Renvoie un tableau contenant toutes les instances Users.
      * Si l'instance $post est fournie, le tableau retourné aura comme premier élément l'auteur de l'instance $post
      */
-    private function authorsList(?Post $post = null) : array
+    private function authorsList(?Post $post = null): array
     {
         if ($post) {
             $usersManager = $this->managers->getManagerOf('User');
@@ -157,13 +179,14 @@ class PostsController extends BackController
             return $authors;
         }
     }
+
     public function processForm(HTTPRequest $request)
     {
         if ($request->method() == 'POST') {
             if (!empty($_FILES['image'])) {
                 if ($_FILES['image']['error']) {
                     $image = $_POST['oldValue'];
-                    if ($_FILES['image']['error']==1) {
+                    if ($_FILES['image']['error'] == 1) {
                         $this->app->user()->setFlash('L\'image ne doit pas dépasser 2 Mo !', 'alert alert-danger');
                     }
                 } else {
@@ -187,13 +210,15 @@ class PostsController extends BackController
 
             if ($request->getExists('id')) {
                 $idPost = $request->getData('id');
-                $post -> setId($idPost);
+                $post->setId($idPost);
+                $authors = $this->authorsList($post);
+                $post->setAuteur($authors);
 
-                if ($_FILES['image']['error']) {
-                    $authors = $this->authorsList($post);
-                    $post->setAuteur($authors);
-                }
-            } elseif ($_FILES['image']['error']) {
+                // if ($_FILES['image']['error']) {
+                //     $authors = $this->authorsList($post);
+                //     $post->setAuteur($authors);
+                // }
+            } else {
                 $authors = $this->authorsList();
                 $post->setAuteur($authors);
             }
@@ -215,7 +240,7 @@ class PostsController extends BackController
 
         $form = $formBuilder->form();
 
-        $formHandler = new FormHandler($form, $this->managers->getManagerOf('Post'), $request);
+        $formHandler = new FormHandler($form, $this->managers->getManagerOf('Post'), $request, $this->app->user());
 
         if (isset($_FILES['image']) and $_FILES['image']['error'] == 1) {
             $this->page->addVar('form', $form->createView());
@@ -230,5 +255,45 @@ class PostsController extends BackController
         }
 
         $this->page->addVar('form', $form->createView());
+    }
+
+    public function executeSeeInvalidComments(HTTPRequest $request)
+    {
+        $this->page->addVar('title', 'Validation des commentaires');
+        $nbInvalidComments = $this->managers->getManagerOf('Comment')->countInvalidComments();
+
+        $invalidComments = $this->managers->getManagerOf('Comment')->getListOfInvalid();
+
+        foreach ($invalidComments as $comment) {
+            $idArticle = $comment->idArticle();
+            $idAuteur = $comment->idAuteur();
+
+            $article = $this->managers->getManagerOf('Post')->getUnique($idArticle);
+            $uriArticle = $this->app->router()->generateUri("showPost", [$article->slug(), $idArticle]);
+            $urlArticle = '<a href="' . $uriArticle . '">' . $article->titre() . '</a>';
+
+            $auteurComment = $this->managers->getManagerOf('User')->getUnique($idAuteur);
+            if ($auteurComment) {
+                $auteurComment = $auteurComment->nom() . ' ' . $auteurComment->prenom() . ' (pseudo : ' . $auteurComment->pseudo() . ')';
+            }
+
+            $comment->urlArticle = $urlArticle;
+            $comment->auteurComment = $auteurComment;
+        }
+
+        $this->page->addVar('nbInvalidComments', $nbInvalidComments);
+        $this->page->addVar('invalidComments', $invalidComments);
+        $this->addView();
+    }
+
+    public function executeValidComment(HTTPRequest $request)
+    {
+        $commentId = $request->getData('id');
+
+        $this->managers->getManagerOf('Comment')->validComment($commentId);
+
+        $this->app->user()->setFlash('Le commentaire a bien été validé !', "alert alert-success");
+
+        $this->app->httpResponse()->redirect('/admin/comments.html');
     }
 }
